fix: extract session storage utils to avoid pulling @lookout/react into SSR
